+++ app/Widgets/Chats/Chats.php
@@ -11,7 +11,7 @@ use App\OpenChat;
 use App\Roster;
 use App\User;
 use Carbon\Carbon;
-use Movim\CurrentCall;
+use Movim\CurrentCalls;
 
 use React\Promise\PromiseInterface;
 use function React\Promise\resolve;
@@ -54,12 +54,21 @@ class Chats extends Base
 
         $this->registerEvent('currentcall_started', 'onCallEvent', 'chat');
         $this->registerEvent('currentcall_stopped', 'onCallEvent', 'chat');
+
+        $this->registerEvent('callinvitepropose', 'onCallInvite');
+        $this->registerEvent('callinviteaccept', 'onCallInvite');
+        $this->registerEvent('callinviteleft', 'onCallInvite');
     }
 
     public function onStart($packet)
     {
         $tpl = $this->tpl();
         $tpl->cacheClear('_chats_item');
+    }
+
+    public function onCallInvite($packet)
+    {
+        $this->rpc('MovimTpl.fill', '#chats_calls_list', $this->prepareCalls());
     }
 
     public function onMessage($packet)
@@ -253,6 +262,14 @@ class Chats extends Base
         }
     }
 
+    public function prepareCalls()
+    {
+        $view = $this->tpl();
+        $view->assign('calls', $this->user->session->mujiCalls()->where('isfromconference', false)->get());
+
+        return $view->draw('_chats_calls');
+    }
+
     public function prepareChats()
     {
         $chats = $this->resolveChats();
@@ -354,7 +371,7 @@ class Chats extends Base
         $view->assign('contact', $contact);
         $view->assign('roster', $roster);
         $view->assign('count', $this->user->unreads($jid));
-        $view->assign('contactincall', CurrentCall::getInstance()->isJidInCall($jid));
+        $view->assign('contactincall', CurrentCalls::getInstance()->isJidInCall($jid));
 
         if ($status == null) {
             $view->assign('message', $message);