+++ src/Movim/Librairies/JingletoSDP.php
@@ -2,13 +2,14 @@
 
 namespace Movim\Librairies;
 
-use Movim\CurrentCall;
+use Movim\CurrentCalls;
 use SimpleXMLElement;
 
 class JingletoSDP
 {
     private string $sdp = '';
     private SimpleXMLElement $jingle;
+    private ?string $sid = null;
 
     private string $action;
 
@@ -33,15 +34,10 @@ class JingletoSDP
         $this->jingle = $jingle;
 
         if (isset($this->jingle->attributes()->sid)) {
-            CurrentCall::getInstance()->id = (string)$this->jingle->attributes()->sid;
+            $this->sid = (string)$this->jingle->attributes()->sid;
         }
 
         $this->action = (string)$this->jingle->attributes()->action;
-    }
-
-    public function getSessionId()
-    {
-        return substr(base_convert(hash('sha256', CurrentCall::getInstance()->id), 30, 10), 0, 6);
     }
 
     public function generate()
@@ -53,7 +49,7 @@ class JingletoSDP
             $username = '-';
         }
 
-        $this->values['session_sdp_id'] = $this->getSessionId();
+        $this->values['session_sdp_id'] = substr(base_convert(hash('sha256', $this->sid), 30, 10), 0, 6);
 
         $sdpVersion =
             'v=0';