+++ src/Moxl/Xec/Payload/JingleReject.php
@@ -2,14 +2,14 @@
 
 namespace Moxl\Xec\Payload;
 
-use Movim\CurrentCall;
+use Movim\CurrentCalls;
 
 class JingleReject extends Payload
 {
     public function handle(?\SimpleXMLElement $stanza = null, ?\SimpleXMLElement $parent = null)
     {
         // We can only reject the current session
-        $jingleSid = CurrentCall::getInstance()->id;
+        $jingleSid = CurrentCalls::getInstance()->id;
         if ($jingleSid && (string)$stanza->attributes()->id != $jingleSid) return;
 
         $this->pack([